requirements.txt
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
Django==3.2
django-mirage-field==1.1.8
django-phonenumber-field==5.1.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
phonenumbers==8.12.21
pycparser==2.20
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
tqdm==4.60.0
